Play nicely with vim's alternate file
Update commands to use keepalt where appropriate such that after opening
and selecting a file from the MRU window, the alternate file is the last
edited file. Works whether MRU opens in current window or a new window.

Note: other plugins, notably netrw, make this an opt in behaviour, but
as MRU files loads blank when you re-open the buffer without running the
MRU command, it doesn't seems necessary to make this an optional setting.